Read all videos from mysql in streaming mode
All the videos were being buffered despite readAll() returning a stream.
The mySql driver needs to be instructed to stream results:

According to
https://dev.mysql.com/doc/connector-j/5.1/en/connector-j-reference-implementation-notes.html

"The combination of a forward-only, read-only result set, with a fetch
size of Integer.MIN_VALUE serves as a signal to the driver to stream
result sets row-by-row. After this, any result sets created with the
statement will be retrieved row-by-row."

Signed-off-by: Jacek Rzeniewicz <jacek@boclips.com>