Increase timeout on ResponseBodyEmitter
Turns out the default (null) doesn't mean no timeout - it
just uses a value from elsewhere in the framework.

Explicitly setting the timeout to 1 hour should allow us
to monitor the progress of long running jobs (such as re-indexing).

Also updated the ResponseEmitterProgressNotifier to catch the exception
arising when trying to send a message to an already closed
ResponseBodyEmitter. It's better to be able to let the job continue
without status updates (logging is still available).

[#162643123]