Update content partners in videos by id not name
When updating a content partner, we need to cascade the update to the
content partner documents inside the videos.

Until today, this was done by finding content partners in videos
matching by name. This is brittle as the name is used as a key. This
commit changes that behaviour, and looks up content partner by id
instead.