VideoSearchUpdater to update search index event-based
This simplifies the logic to keep the index up-to-date dramatically.
Whenever a video is updated (title, distribution methods, etc.) the
search index will be updated automatically.

Future optimisations are batch updates, and preventing updates which
are already in the right state.

Upon VideoUpdated / VideoCreated we look up the video in the database.
This can be seen as a bad pattern, as other data mutation can have
happened in the meanwhile. Further discussions are necessary.