Allow retrieving a specific collection by id through contracts
Enabling this required a bit of refactoring. Previously we had Kotlin
extension functions attached to collectionRepository that did the ACL
business.

To add the logic for verifing contracts I needed to inject a new component
over there, but the extension functions approach didn't allow it in a
clean way (I'd probably need to inject into a Mongo repository and
cast...)

I moved those extension functions as methods of CollectionService
and replaced instances of repository with the service when it was
necessary.

[#168135330]