Keycloak client to not depend on external modules
ObjectMapperDefinition is referencing Jackson2Hal module which is moving
package in spring-hateoas 1.0. Once video-service client has been
upgraded, we can ship the dependency with the client. Right now the
dependency in the client is older and is replaced with a newer version
on the consumer side. In that newer version, the import path for
Jackson2Hal is incorrect and it yields a NoClassDefFound.