Owner & bookmarked search filters are inclusive
We considered adding a new endpoint /users/{userId}/collections as a way
of accessing a search that returns both a users own collections, and the
users bookmarked collections.

However, in order to do this we would have needed to use the underlying
search query that the /collections endpoint uses anyway.

We appreciate that this change fundamentally changes the symantics of
the owner and bookmarked filters on the API by changing them to an OR
from the implicit AND that was there before.

Without huge rethinking of the API / search query this is probably the
best solution for now.

[#170072747]

Signed-off-by: Antony O'Neill <antony@boclips.com>