Expose suggestions endpoint
For now the suggestions endpoint serves content partner name suggestions
for an 3 <= input <= 20.

The content partners are looked up in the database, we're intending this
to change to ES instead.

We envision this to become the main endpoint to power autocomplete
search.

Some big questions:
* How do we make tracking work - to measure the impact of autocomplete
and associate searches with suggestions
* What information are we actually interested in suggestions - super
sparse like name, or more?

[#171896867]

Signed-off-by: Benjamin Gröhbiel <ben@boclips.com>