Refactor elasticsearch query steps
We will apply the access rule filters in the beginning, so when it does
the aggregations, it does not have to unneccessarily process all the
documents. User provided filters will still be applied at the very end,
because we need this to be able to calculate the proper facet counts.

[#173528062]