Video can be accessed with shareCode
We don't want to send playback & attachment details to anonymous users
who are not providing a valid share code. As a first step, if they don't
provide shareCode and referer we return the full video. If they do
provide it, we validate it and if it's invalid we omit playback &
attachments. Once teachers is updated to send these details, we can
update VS to always require the code from anonym users for the full video access.

[#173916773]