Attempt to fix video search slowness
We think the problem was to do with channel facet hydration. It looks
like we were loading all channels into memory, to then filter them out
by ID. A more optimised approach would be to do the filtering in mongo.

We're still not sure why loading all channels into memory was causing
flaky performance. But hopefully this will fix the slowness

[#176149515]

Signed-off-by: Alex Glen <alex@boclips.com>