Fix ES issue with null prices
The script was throwing due to null prices in e2e, this seems to
replicate and fix the scenario

I've also removed the guard for a user needing an organisation whilst
searching. We can add this back in once our user domain enforces this
properly. There are still some users without an org

[#175854287]