Fix price aggregation for when no prices returned
We were seeing `No field found for [prices.6013076fe00f31001d5cc989] in
mapping with types` in the testing environment. This adds a test that
reproduces this issue. Right now if there is a set of search results
that have no prices on the videos, we'll get a facet with a price of 0.
We would probably want to filter this out instead of returning 0, maybe.

[#175854287]