Refactor more user service calls into the User
- These userService calls felt a bit scattered all over the place,
especially as we already have an abstraction for this.
- Moved prices/organisation/canViewPrices into the User domain

This also has a nice advantage of being able to use the lazy keyword,
and so can essentially cache this data between requests as this user
gets created in the base controller. Though, maybe this lazy business doesn't quite belong in the user domain

[#177740625]