Prepare ChannelsIndexReader for basic search
Refactoring around ChannelsIndexReader so it not only supports getting
suggestions but can be used to actually search for channels, sort them,
paginate etc. This was mostly renaming and adding more types, no
functionality changed. There is probably a simpler way to do this but I got
entangled in it too many times now, so going with this approach for the
time being.

[#177933844]